+++ src/index.js
@@ -32,6 +32,11 @@ import CompleteDataSetButton from "./components/dataentry/CompleteDataSetButton"
 import { generateCalculator } from "./components/dataentry/CalculatorFactory";
 import HesabuFormula from "./components/dataentry/HesabuFormula";
 import EditIasoFormButton from "./components/dataentry/EditIasoFormButton";
+import { toOrgUnitFacts } from "./components/dataentry/CodeGenerator"
+import DecisionTable from "./components/dataentry/DecisionTable"
+import { getVisibleAndOrderedActivities } from "./components/dataentry/getVisibleAndOrderedActivities"
+
+import IasoIcon from "./components/shared/icons/IasoIcon";
 
 const App = AppDrawer
 
@@ -66,5 +71,10 @@ export {
   CompleteDataSetButton,
   generateCalculator,
   HesabuFormula,
-  EditIasoFormButton
+  EditIasoFormButton,
+  DecisionTable,
+  toOrgUnitFacts,
+  getVisibleAndOrderedActivities,
+  // icons
+  IasoIcon,
 };